Removed conversion warning.

+++ src/Bpp/Numeric/Stat/ContingencyTableTest.cpp
@@ -67,7 +67,7 @@ ContingencyTableTest::ContingencyTableTest(const std::vector<std::vector<size_t>
   {
     for (size_t j = 0; j < m; ++j)
     {
-      long double c = table[i][j];
+      long double c = static_cast<long double>(table[i][j]);
       long double e = static_cast<long double>(margin1_[i] * margin2_[j]) / static_cast<long double>(tot);
       expc(i, j) = e;
       statistic_ += static_cast<double>(std::pow(c - e, 2.L) / e);
@@ -88,7 +88,7 @@ ContingencyTableTest::ContingencyTableTest(const std::vector<std::vector<size_t>
       {
         for (size_t j = 0; j < m; ++j)
         {
-          long double c = table_rep(i, j);
+          long double c = static_cast<long double>(table_rep(i, j));
           long double e = expc(i, j);
           stat_rep += static_cast<double>(std::pow(c - e, 2.L) / e);
         }
